Add files via upload
Search bar created from HTML with CSS styling addition. Search bar takes the search request and redirects to the existing search page with the query. An attempt was made to match the style of the website. For this commit, time was spent learning basic HTML, CSS,-albeit ultimately unsuccessful-PHP, Django, and Wagtail. Additions to this feature are being actively considered.