Home page blocks (#169)
* WIP: Implement override image block.

Implemented an override image block but it currently
only works for the single column. The single
column behaves completely differently from the
two columns and three columns. I will revist
this.

* WIP: No image center div

When there is no image, center the div
and span the entire row. Ideally give options.

* fix: Implement image override for multiple columns

Implemented the image override feature for all column
types. This was done by passing an image value to
the child elements.

* chore: Remove todo
Removed a todo message.

* fix: Add override headline to template

Fixed the issue of no override ability
for titles in headlines. The template
was missing the switches on the headlines.