Searchbar (#170)
* Add files via upload

Search bar created from HTML with CSS styling addition. Search bar takes the search request and redirects to the existing search page with the query. An attempt was made to match the style of the website. For this commit, time was spent learning basic HTML, CSS,-albeit ultimately unsuccessful-PHP, Django, and Wagtail. Additions to this feature are being actively considered.

* Add files via upload

Search bar created from HTML with CSS styling addition. Search bar takes the search request and redirects to the existing search page with the query. An attempt was made to match the style of the website. For this commit, time was spent learning basic HTML, CSS,-albeit ultimately unsuccessful-PHP, Django, and Wagtail. Additions to this feature are being actively considered.

* Add files via upload

Added to support mobile devices. Searchbar is hidden in favor for search icon for displays that do not have the real-estate to render it well.

* Add files via upload

Added to support mobile devices. Searchbar is hidden in favor for search icon for displays that do not have the real-estate to render it well.

* Add files via upload

Test of Using JQuery autofill feature. Note: Attempted to use HTML's list for autocomplete, but I found that it is less flexible than JQuery.

* Add files via upload

Test of Using JQuery autofill feature. Note: Attempted to use HTML's list for autocomplete, but I found that it is less flexible than JQuery.

* Add files via upload

Changes: Fixed URL redirect, fixed bug with it breaking WebKit, and completed JQuery Searchbar Autofill supporting multiple entries which references a custom made array storing commonly searched terms. Final Update before pull request.

* Add files via upload

* Update to Readme: Ubuntu

Updated instructions such that user only needs to follow them exactly to get pipeline up and running on their Ubuntu system.